Move non-models to tictactoe/game.py
